Introduce `Signature` to scan for memory signatures
Memory signatures are often used for finding functions within libraries,
by describing the way parts of the function look, usually starting with
the prologue.

You can use `?` to represent "any value" to be acceptable, which is
useful for dealing with offsets within functions within binaries that
are re-compiled, but without functionally different code.