Revert 480338a and use `int64_t` when doing math with `size_t`
I misread the poorly-formatted manpage for `ssize_t`, it is not in
`stddef.h` but rather `sys/types.h` -- which Windows of course does not
have.

A `SSIZE_T` macro exists, and should probably be used for a typedef in
the future.

For now, I'm just going to do all the math inside an `int64_t` and hope
it's big enough.